+++ cc-service-api/src/main/java/com/jtmc/apps/reforma/api/v1/persona/PersonaResponse.java
@@ -3,7 +3,34 @@ package com.jtmc.apps.reforma.api.v1.persona;
 public class PersonaResponse {
 
     private int id;
-    private String completeName;
+    private String name;
+    private String lastname;
+    private String registration;
+    private  boolean status;
+
+    public String getLastname() {
+        return lastname;
+    }
+
+    public void setLastname(String lastname) {
+        this.lastname = lastname;
+    }
+
+    public String getRegistration() {
+        return registration;
+    }
+
+    public void setRegistration(String registration) {
+        this.registration = registration;
+    }
+
+    public boolean isStatus() {
+        return status;
+    }
+
+    public void setStatus(boolean status) {
+        this.status = status;
+    }
 
     public int getId() {
         return id;
@@ -13,11 +40,11 @@ public class PersonaResponse {
         this.id = id;
     }
 
-    public String getCompleteName() {
-        return completeName;
+    public String getName() {
+        return name;
     }
 
-    public void setCompleteName(String completeName) {
-        this.completeName = completeName;
+    public void setName(String name) {
+        this.name = name;
     }
 }