+++ cc-service-impl/src/main/java/com/jtmc/apps/reforma/impl/branch/BranchImpl.java
@@ -2,6 +2,7 @@ package com.jtmc.apps.reforma.impl.branch;
 
 import com.google.inject.Inject;
 import com.jtmc.apps.reforma.api.v1.branch.BranchInitialAmount;
+import com.jtmc.apps.reforma.api.v1.branch.BranchRequest;
 import com.jtmc.apps.reforma.domain.*;
 import com.jtmc.apps.reforma.impl.catalogcount.CatalogCountImpl;
 import com.jtmc.apps.reforma.impl.user.UserImpl;
@@ -57,15 +58,23 @@ public class BranchImpl {
     }
 
     public void insertBranch(Branch branch) {
-        userImpl.getLoggedInUserDetails();
+        userImpl.validateAdminPermissionsForLoggedInUser();
         if (branchRepository.insertBranch(branch) != 1) {
             logger.error("Branch was not inserted. {}", branch);
             throw new RuntimeException("Branch was not inserted.");
         }
     }
 
-    public void insertInitialAmount(Branch branch, double amount) {
+    public void updateBranch(Branch branch) {
+        userImpl.validateAdminPermissionsForLoggedInUser();
+        if (branchRepository.updateBranch(branch) != 1) {
+            logger.error("Branch was not updated. {}", branch);
+            throw new RuntimeException("Branch was not updated.");
+        }
+    }
+
+    public void insertInitialAmount(BranchDetails branchDetails, double amount) {
         UserDetails userDetails = userImpl.validateWritePermissionsForLoggedInUser();
-        catalogCountImpl.insertInitialAmountCatalogCount(branch, amount);
+        catalogCountImpl.insertInitialAmountCatalogCount(branchDetails, amount);
     }
 }